refactor: 가독성과 관리 용이성을 위해 디렉토리 구조 변경
- 테스트 디렉토리 포함
