+++ src/main/java/racingcar/viewer/RacingCarViewer.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.viewer;
 
 import java.util.List;
 import java.util.Map;
